requirements.txt
# Core dependencies
Flask==3.0.0
Pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78  # headless версия для сервера
rembg==2.0.52
onnxruntime==1.16.3
gunicorn==21.2.0  # Production WSGI сервер
requests==2.31.0